third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-window-name-navigator-expected.txt

ALERT: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
ALERT: PASS
ALERT: PASS
ALERT: PASS
ALERT: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame.
 
